File to redirect everything to 80 except a few pages
When you are on a SSL page, say https://domainname.com/ssl_page1, then
if you click any links inside that page, then you'd be clicking the
HTTPS part of those links. So, there needs to be a way to make them to
be redirected to port 80. This code does just that.